Fixing Jamuma #466
The issue is that Client->dup_check doesn't return Client objects, it returns a deconstructed hash ref of select bits of the client.  But the Venus templates still think its a full client.

This is for client permissions reasons: you might not have permissions to view the dupliciate clients.  So it returns a cleansed hash ref of innocent data.  For this reason, ssn has not been restored even though the Venus templates want it.

I also removed the cutesy "pronoun" logic from the Venus templates as it would necessitate leaking the client's gender.  Probably not a big deal, but the pronoun logic isn't worth it.

This was broken doing emc #1254 by permissions@7573.