Add emc::Rolodex::ByState and accompanying HasByState role
This is for attaching per-state information to a rolodex entry.  Currently the only
per state information necessary is the ID number of a doctor's license to practice in that
state.  I figure more will be coming.

Yes, this will only work for US and maybe Canada and Mexico.  But EMC has a lot of other issues
to be solved before it can be used internationally.

Rather than adding to Rolodex.pm I wrote the methods to access ByState in a role, mostly
for code organization reasons.

min_schema_required() doesn't do anything, maybe it will in the future, seemed better than
a comment.