Remove the run_report_pdf() stub.
The plan was to overlay the existing PDFs with client and treater data.  This proved too
involved to get the report system to do, and too much maint when the PDFs change.  Instead
a web monkey will convert the PDFs to HTML and we'll make regular reports out of them.

I'll leave in the ability for a plugin to specify an op, it will probably prove useful in
the future.