Turn the CA specific Attenting Physician Statement from Jamuma into a template.
This is enough to demonstrate that it works and looks right.  Checking in to get a clean slate.

The WA template is also checked in but not made into a template.

The attending phys report does not pick the template by state yet.

The Client Intake Forms report is messed up with double print headers.